# public/knowledge-api-openapi.yaml
openapi: 3.0.3
info:
  title: Knowledge Management API
  description: |
    RESTful API for the knowledge management system.
    
    ## Overview
    This API provides endpoints for:
    - **Search** for full-text and semantic search across knowledge articles
    - **Articles** for article CRUD operations
    - **Analytics** - for search queries, results, agent interactions, and usage metrics
    
    ## Authentication
    All endpoints require authentication via JWT Bearer tokens obtained through OAuth 2.0 / OpenID Connect.
    
    ## Rate Limiting
    - Standard tier: 100 requests/minute per user
    - Professional tier: 1,000 requests/minute per user
    - Enterprise tier: Custom limits
    
    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the window resets
    
    ## Versioning
    API versioning is done via URL path (`/api/v1/`). Breaking changes will increment the major version.
    Deprecation notices are provided via `Sunset` header 6 months before removal.

  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: Proprietary
...
security:
  - bearerAuth: []
paths:
...
  /search:
    post:
      tags:
        - Search
      summary: Search knowledge articles
      description: |
        Performs full-text and semantic search across knowledge articles.
        
        **Search Features (assuming the proposed architecture):**
        - Keyword matching with BM25 algorithm
        - Semantic search using vector embeddings (kNN)
        - Hybrid search combining both approaches
        - Faceted filtering by category, date, tags
        
        **Performance:**
        - Results cached for 5 minutes
        - p95 latency target: <100ms for cached, <500ms for uncached

      operationId: searchArticles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            examples:
              simpleSearch:
                summary: Simple keyword search
                value:
                  query: "refund process"
                  page: 1
                  pageSize: 10
              filteredSearch:
                summary: Search with filters
                value:
                  query: "password reset"
                  category: "Technical Support"
                  dateFrom: "2025-01-01"
                  dateTo: "2025-12-31"
                  sortBy: "relevance"
                  page: 1
                  pageSize: 20
              semanticSearch:
                summary: Semantic/AI search
                value:
                  query: "customer is angry about late delivery"
                  searchType: "semantic"
                  includeAiSummary: true
      responses:
        '200':
          description: Search results
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            X-Cache:
              schema:
                type: string
                enum: [HIT, MISS]
              description: Cache status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
              example:
                articles:
                  - id: "art_123"
                    title: "How to Process Customer Refunds"
                    snippet: "This article explains the step-by-step process for handling customer refund requests..."
                    category: "Billing"
                    tags: ["refund", "billing", "returns"]
                    relevanceScore: 0.95
                    lastUpdated: "2024-12-01T14:30:00Z"
                    viewCount: 1542
                total: 45
                page: 1
                pageSize: 10
                hasMore: true
                searchId: "srch_abc123"
                processingTimeMs: 45
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /search/suggestions:
    get:
      tags:
        - Search
      summary: Get search suggestions
      description: |
        Returns autocomplete suggestions based on partial query input.
        Combines:
        - Popular searches
        - User's recent searches
        - AI-suggested completions
      operationId: getSearchSuggestions
      parameters:
        - name: q
          in: query
          required: true
          description: Partial search query (min 2 characters)
          schema:
            type: string
            minLength: 2
            maxLength: 100
          example: "refun"
        - name: limit
          in: query
          required: false
          description: Maximum suggestions to return
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 5
      responses:
        '200':
          description: Search suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuggestionsResponse'
              example:
                suggestions:
                  - text: "refund process"
                    type: "popular"
                    count: 1250
                  - text: "refund policy"
                    type: "popular"
                    count: 890
                  - text: "refund status check"
                    type: "recent"
                  - text: "refund for damaged item"
                    type: "suggested"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================================
  # ARTICLE CRUD ENDPOINTS
  # ============================================================================
...
  /analytics/search:
    get:
      tags:
        - Analytics
      summary: Get search analytics
      description: |
        Retrieves aggregated search analytics for reporting.
        
        **Authorization:** Requires `analytics:read` scope.
        
        **Available Metrics:**
        - Total searches
        - Unique searchers
        - Zero-result searches
        - Average results per search
        - Click-through rate
        - Top queries
        - Search latency percentiles
      operationId: getSearchAnalytics
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2024-12-01"
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2024-12-19"
        - name: granularity
          in: query
          description: Time bucket granularity
          schema:
            type: string
            enum: [hour, day, week, month]
            default: day
        - name: metrics
          in: query
          description: Metrics to include (comma-separated)
          schema:
            type: string
          example: "totalSearches,uniqueUsers,ctr"
      responses:
        '200':
          description: Search analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchAnalyticsResponse'
              example:
                period:
                  start: "2024-12-01"
                  end: "2024-12-19"
                  granularity: "day"
                summary:
                  totalSearches: 45230
                  uniqueUsers: 1250
                  avgResultsPerSearch: 8.5
                  zeroResultRate: 0.03
                  clickThroughRate: 0.72
                  avgLatencyMs: 85
                timeSeries:
                  - date: "2024-12-01"
                    searches: 2100
                    uniqueUsers: 320
                  - date: "2024-12-02"
                    searches: 2450
                    uniqueUsers: 345
                topQueries:
                  - query: "refund"
                    count: 1523
                    ctr: 0.85
                  - query: "password reset"
                    count: 1201
                    ctr: 0.78
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained via OAuth 2.0 / OpenID Connect.
        
        **Token Lifetime:** 1 hour
        
        **Required Claims:**
        - `sub` - User ID
        - `scope` - Permission scopes
        - `tenant_id` - Multi-tenant identifier
        
        **Example:**
        ```
        Authorization: Bearer eyJhbGciOiJSUzI1NiIs...
        ```

  # Reusable Parameters
...
  headers:
    X-RateLimit-Limit:
      description: Maximum requests per time window
      schema:
        type: integer
      example: 100
    X-RateLimit-Remaining:
      description: Remaining requests in current window
      schema:
        type: integer
      example: 95
    X-RateLimit-Reset:
      description: Unix timestamp when window resets
      schema:
        type: integer
      example: 1703001600

  # Schemas
  schemas:
...
    SearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 500
          description: Search query string
        searchType:
          type: string
          enum: [keyword, semantic, hybrid]
          default: hybrid
          description: Type of search to perform
        category:
          type: string
          description: Filter by category
        tags:
          type: array
          items:
            type: string
          description: Filter by tags (OR logic)
        dateFrom:
          type: string
          format: date
          description: Filter articles created after this date
        dateTo:
          type: string
          format: date
          description: Filter articles created before this date
        sortBy:
          type: string
          enum: [relevance, date, popularity]
          default: relevance
        page:
          type: integer
          minimum: 1
          default: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
        includeAiSummary:
          type: boolean
          default: false
          description: Include AI-generated answer summary
    SearchResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: '#/components/schemas/ArticleSummary'
        total:
          type: integer
          description: Total matching articles
        page:
          type: integer
        pageSize:
          type: integer
        hasMore:
          type: boolean
        searchId:
          type: string
          description: Unique ID for analytics tracking
        processingTimeMs:
          type: integer
          description: Server processing time
        aiSummary:
          type: string
          description: AI-generated answer (if requested)
    SuggestionsResponse:
      type: object
      properties:
        suggestions:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              type:
                type: string
                enum: [popular, recent, suggested]
              count:
                type: integer
                description: Search count (for popular suggestions)

    # Article Schemas
...
    ArticleSummary:
      type: object
      description: Abbreviated article for search results
      properties:
        id:
          type: string
        title:
          type: string
        snippet:
          type: string
          description: Highlighted content excerpt
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        relevanceScore:
          type: number
        lastUpdated:
          type: string
          format: date-time
        viewCount:
          type: integer
...
    SearchAnalyticsResponse:
      type: object
      properties:
        period:
          type: object
          properties:
            start:
              type: string
              format: date
            end:
              type: string
              format: date
            granularity:
              type: string
        summary:
          type: object
          properties:
            totalSearches:
              type: integer
            uniqueUsers:
              type: integer
            avgResultsPerSearch:
              type: number
            zeroResultRate:
              type: number
            clickThroughRate:
              type: number
            avgLatencyMs:
              type: integer
        timeSeries:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              searches:
                type: integer
              uniqueUsers:
                type: integer
        topQueries:
          type: array
          items:
            type: object
            properties:
              query:
                type: string
              count:
                type: integer
              ctr:
                type: number
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error context
        correlationId:
          type: string
          description: Request ID for support inquiries
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "BAD_REQUEST"
            message: "Invalid request parameters"
            details:
              invalidParams: ["pageSize must be between 1 and 100"]
            correlationId: "req_xyz789"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "UNAUTHORIZED"
            message: "Authentication required"
            correlationId: "req_xyz789"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "FORBIDDEN"
            message: "Insufficient permissions to access this resource"
            details:
              requiredScope: "articles:write"
            correlationId: "req_xyz789"
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          schema:
            type: integer
          example: 0
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        Retry-After:
          description: Seconds until rate limit resets
          schema:
            type: integer
          example: 30
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "RATE_LIMITED"
            message: "Rate limit exceeded. Please retry after 30 seconds."
            details:
              limit: 100
              window: "1 minute"
              retryAfter: 30
            correlationId: "req_xyz789"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "INTERNAL_ERROR"
            message: "An unexpected error occurred"
            correlationId: "req_xyz789"